+++ src/granularity_convert/CMakeLists.txt
@@ -1,2 +1,4 @@
 target_sources(plateau PRIVATE
-        "granularity_converter.cpp")
+        "granularity_converter.cpp"
+        "node_path.cpp"
+)
